# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "成果云系统API文档",
    "description": "成果云系统后端接口文档（Flask + MySQL）",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "登录后返回的 Bearer Token，格式：Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/api/students/": {
      "get": {
        "tags": ["学生"],
        "summary": "学生列表（分页/搜索）",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回学生列表"}}
      },
      "post": {
        "tags": ["学生"],
        "summary": "添加学生（教师录入直接生效）",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "body", "in": "body", "schema": {"type": "object"}}
        ],
        "responses": {"201": {"description": "添加成功"}}
      }
    },
...
    "/api/student/export/certificates": {
      "post": {
        "tags": ["导出"],
        "summary": "学生导出个人获奖证书",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回证书文件"}}
      }
    },
    "/api/teacher/export/award-types": {
      "get": {
        "tags": ["导出"],
        "summary": "获奖类型字典",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回获奖类型"}}
      }
    },
    "/api/teacher/export/student/certificates": {
      "post": {
        "tags": ["导出"],
        "summary": "教师批量导学生证书",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回证书文件"}}
      }
    },
...
  }
}